adapt tests even more.
